requirements.txt
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typer>=0.9.0 